pyproject.toml
[tool.poetry]
name = "course-work-upit"
version = "0.1.0"
description = ""
authors = ["Aleksandr <zemacohoznik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
